+++ src/carp/range_utils.cc
@@ -34,7 +34,7 @@ void pivot_union(std::vector<rb_item_t> rb_items,
   /* Ranks that currently have an active bin */
   std::list<int> active_ranks;
 
-  for (int i = 0; i < rb_items.size(); i++) {
+  for (size_t i = 0; i < rb_items.size(); i++) {
     int bp_rank = rb_items[i].rank;
     float bp_bin_val = rb_items[i].bin_val;
     float bp_bin_other = rb_items[i].bin_other;
